requirements.txt
streamlit
google-genai
streamlit-js-eval

